refactor: 입출력 검증을 뷰에서 분리 RaceController.readFromConsole 메서드 위치 뷰로 변경
