+++ src/main/java/racing/controller/RacingController.java
@@ -1,15 +1,13 @@
 package racing.controller;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
 import java.util.List;
+import racing.domain.CarNameInputValidator;
 import racing.domain.RacingGame;
+import racing.domain.RacingMaxTurnInputValidator;
 import racing.domain.RandomGenerator;
 import racing.dto.RacingResults;
-import racing.view.CarNameInputView;
 import racing.view.InputGuideOutputView;
-import racing.view.RacingMaxTurnInputView;
+import racing.view.InputView;
 import racing.view.RacingResultOutputView;
 import util.RetryHelper;
 
@@ -27,7 +25,7 @@ public class RacingController {
     RetryHelper retryHelper = new RetryHelper(5);
     return retryHelper.retry(() -> {
       InputGuideOutputView.printInputCarNames();
-      return CarNameInputView.getName(this::readFromConsole);
+      return InputView.getInput(CarNameInputValidator::getName);
     });
   }
 
@@ -35,16 +33,7 @@ public class RacingController {
     RetryHelper retryHelper = new RetryHelper(5);
     return retryHelper.retry(() -> {
       InputGuideOutputView.printInputMaxRacingTurn();
-      return RacingMaxTurnInputView.getMaxTurn(this::readFromConsole);
+      return InputView.getInput(RacingMaxTurnInputValidator::getMaxTurn);
     });
-  }
-
-  private String readFromConsole() {
-    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
-    try {
-      return bufferedReader.readLine();
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
   }
 }
